+++ lib/core/options.py
@@ -761,6 +761,7 @@ def merge_config(opt: Values) -> Values:
     opt.output_formats = opt.output_formats or config.safe_get(
         "output", "output-formats", "plain"
     )
+    opt.save_response = opt.save_response or config.safe_get("output", "save-response")
     opt.log_file = opt.log_file or config.safe_get("output", "log-file")
     opt.log_file_size = config.safe_getint("output", "log-file-size")